requirements.txt
Django==4.2.2
djangorestframework==3.14.0
Python 3.11.1
pip 23.1.2
requests==2.31.0
